+++ ftl/integration_tests/ftl_python_integration_tests_yaml.py
@@ -9,7 +9,7 @@ import util
 TEST_DIRS = [
     'packages_test', 'metadata_test',
     'python3_test', 'pipfile_test',
-    'venv_dir_test', 'no_descriptor_test',
+    'virtualenv_dir_test', 'no_descriptor_test',
     'empty_descriptor_test',
     'additional_directory',
     'commented_descriptor_test',
@@ -46,7 +46,7 @@ def main():
     test_map['metadata_test'].extend(['--exposed-ports', '8090,8091'])
     test_map['python3_test'].extend(['--python-cmd', 'python3.6'])
     test_map['python3_test'].extend(['--pip-cmd', 'python3.6 -m pip'])
-    test_map['venv_dir_test'].extend(['--virtualenv-dir', '/alternate-env'])
+    test_map['virtualenv_dir_test'].extend(['--virtualenv-dir', '/alternate-env'])
     test_map['additional_directory'].extend([
         '--additional-directory',
         '/workspace/ftl/python/testdata/additional_directory'